feat: stock_basic daily

+++ app/models/daily.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, BigInteger, Numeric, String, Date
+from sqlalchemy import Column, Integer, BigInteger, Numeric, String, Date, DECIMAL
 from app.models import db
 
 class Daily(db.Model):
@@ -7,12 +7,12 @@ class Daily(db.Model):
     id = Column(Integer, primary_key=True)
     ts_code = Column(String(10), index=True)
     trade_date = Column(Date)
-    open = Column(Numeric(2))
-    high = Column(Numeric(2))
-    low = Column(Numeric(2))
-    close = Column(Numeric(2))
-    pre_close = Column(Numeric(2))
-    change = Column(Numeric(2))
-    pct_change = Column(Numeric(2))
-    vol = Column(BigInteger)
-    amount = Column(Numeric(2))
+    open = Column(Numeric(6, 2))
+    high = Column(Numeric(6, 2))
+    low = Column(Numeric(6, 2))
+    close = Column(Numeric(6, 2))
+    pre_close = Column(Numeric(6, 2))
+    change = Column(Numeric(6, 2))
+    pct_change = Column(Numeric(6, 4))
+    vol = Column(Numeric(15, 3))
+    amount = Column(Numeric(15, 3))
